refactoring work in progress...
