+++ keyboards/phelix/phelixzero/ui/event_router.h
@@ -1,0 +1,8 @@
+// Copyright 2024 shawnfelix (@shawnfelix)
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#pragma once
+
+#include "globals.h"
+
+void route_btn_event(wui_btn_t btn);
