+++ keyboards/phelix/phelixzero/ui/features/pomodoro.c
@@ -4,8 +4,9 @@
 
 #include <qp.h>
 #include "qp_lvgl.h"
-#include "globals.h"
 #include "pomodoro.h"
+#include "ui_state.h"
+#include "globals.h"
 
 
 void update_pomo_state(pomo_state_t state) {
@@ -30,6 +31,9 @@ void start_pomodoro(void) {
         case POMO_PAUSED:
             wui_gbl_state[WUI_POMODORO].pomo_wui_state.token = defer_exec(1000, update_pomodoro_callback, &wui_gbl_state[WUI_POMODORO]);
             update_pomo_state(POMO_RUNNING);
+            break;
+        case POMO_RUNNING:
+            pause_pomodoro();
             break;
         default:
             break;
@@ -57,9 +61,8 @@ void reset_pomodoro(void) {
     }
 }
 
-void init_widget_pomodoro(void) {
+void init_widget_pomodoro(lv_obj_t *screen) {
     if (wui_gbl_state[WUI_POMODORO].wui_init_state == NOT_INIT) {
-        println("initializing pomodoro....");
         // init pomo wui state in global context
         wui_gbl_state[WUI_POMODORO].pomo_wui_state = (pomo_wui_state_t){
                 .pomo_timer = 0,
@@ -68,7 +71,7 @@ void init_widget_pomodoro(void) {
             };
         wui_gbl_state[WUI_POMODORO].wui_init_state = INIT;
         // init ui stuff
-        wui_gbl_state[WUI_POMODORO].pomo_wui_state.pomo_label = lv_label_create(scr_home);
+        wui_gbl_state[WUI_POMODORO].pomo_wui_state.pomo_label = lv_label_create(screen);
         lv_label_set_text_fmt(wui_gbl_state[WUI_POMODORO].pomo_wui_state.pomo_label, "%02d:%02d:%02d", 0, 0, 0);
         lv_obj_set_style_text_color(wui_gbl_state[WUI_POMODORO].pomo_wui_state.pomo_label, lv_color_make(163, 169, 194), LV_PART_MAIN);
         lv_obj_set_style_text_font(wui_gbl_state[WUI_POMODORO].pomo_wui_state.pomo_label, &lv_font_montserrat_38, LV_PART_MAIN);
@@ -93,12 +96,10 @@ void do_btn_action_pomodoro(wui_btn_t btn) {
 }
 
 void ui_show_pomodoro(void) {
-    println("showing pomo");
     lv_obj_clear_flag(wui_gbl_state[WUI_POMODORO].pomo_wui_state.pomo_label,LV_OBJ_FLAG_HIDDEN);
     lv_obj_fade_in(wui_gbl_state[WUI_POMODORO].pomo_wui_state.pomo_label, 400, 100);
 }
 
 void ui_hide_pomodoro(void) {
-    println("hiding pomo");
     lv_obj_add_flag(wui_gbl_state[WUI_POMODORO].pomo_wui_state.pomo_label,LV_OBJ_FLAG_HIDDEN);
 }
