+++ keyboards/phelix/phelixzero/ui/globals.c
@@ -8,10 +8,9 @@
 #include "features/pk_wpm.h"
 #include "controls.h"
 #include "native/native.h"
+#include "common_types.h"
+#include "features/features.h"
 
-wui_gbl_state_t wui_gbl_state[NumberOfWuiTypes];
-gbl_ui_state_t gbl_ui_state;
-lv_obj_t *scr_home;
 
 const char* enumToString(wui_t wui_enum) {
     switch (wui_enum) {
@@ -25,17 +24,12 @@ void register_wuis(void) {
     // doing this manually sucks
     wui_gbl_state[WUI_POMODORO].fun_btn_event = &do_btn_action_pomodoro;
     wui_gbl_state[WUI_WPM].fun_btn_event = &do_btn_action_pomodoro;
-
 }
 
 void init_all_wuis(void) {
     wui_gbl_state[WUI_POMODORO].wui_init_state = NOT_INIT;
     init_widget_pomodoro();
     register_wuis();
-}
-
-void init_global_ui_state(void) {
-    gbl_ui_state = (gbl_ui_state_t) { .active_layer_text="Default", NULL, false, ""};
 }
 
 
@@ -67,21 +61,3 @@ void minimize_wui(wui_t wui_enum) {
 
 }
 
-/* routes ui input to the active feature */
-//TODO probably move this to its own module
-void btn_event_router(wui_btn_t btn) {
-    if (btn == B5) {
-        toggle_native_elements_visibility();
-        return;
-    }
-
-    // TODO replace this check with a global UI mode (native vs wui/child)
-    if (gbl_ui_state.widget_selector_mode_active == true) {
-        // if the widget selector is open
-        println("global router with widget selector mode");
-        native_btn_event_router(btn);
-    } else {
-        // TODO wui btn event router
-        wui_gbl_state[gbl_ui_state.curr_active_wui].fun_btn_event(btn);
-    }
-}
