+++ keyboards/phelix/phelixzero/ui/globals.h
@@ -3,21 +3,19 @@
 #pragma once
 
 #include "ui_state.h"
-#include "features/features.h"
+#include "native/native.h"
+#include "common_types.h"
+#include "features.h"
 
-// wui_gbl_state_t stores the state of all WUI widgets
 extern wui_gbl_state_t wui_gbl_state[NumberOfWuiTypes];
+extern native_t n_gbl_state[NumberOfNativeComponents];
 extern gbl_ui_state_t gbl_ui_state;
-extern lv_obj_t *scr_home;
 
 const char* enumToString(wui_t wui_enum);
-
-void init_global_ui_state(void);
 
 void set_wui_window_state(wui_t wui_enum, window_state_t state);
 
 void set_gbl_wui(wui_t wui_enum, window_state_t init_state, window_state_t ui_state);
 
-void btn_event_router(wui_btn_t btn);
 
 void init_all_wuis(void);
