+++ keyboards/phelix/phelixzero/ui/native/scroll_selector.h
@@ -2,6 +2,11 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 #pragma once
 #include "native_defs.h"
+#include "controls.h"
+
+typedef struct {
+    bool placeholder; // TODO
+} scroll_selector_t;
 
 void init_ui_widget_selector(lv_obj_t * screen, uint32_t number_of_init_wuis);
