+++ keyboards/phelix/phelixzero/ui/theme.h
@@ -1,0 +1,19 @@
+// Copyright 2024 shawnfelix (@shawnfelix)
+// SPDX-License-Identifier: GPL-3.0-or-later
+#pragma once
+#include <lv_color.h>
+
+extern lv_color_t THEME_ACCENT;
+extern lv_color_t THEME_PRIMARY_BG;
+extern lv_color_t THEME_SECONDARY_BG;
+extern lv_color_t THEME_TEXT_DARK;
+extern lv_color_t THEME_TEXT_LIGHT;
+extern lv_color_t THEME_TEXT_LIGHT_GRAY;
+// UI action buttons
+extern lv_color_t THEME_BUTTON_1;
+extern lv_color_t THEME_BUTTON_2;
+extern lv_color_t THEME_BUTTON_3;
+extern lv_color_t THEME_BUTTON_4;
+
+
+void init_theme_colors(void);
