+++ keyboards/phelix/phelixzero/ui/ui.c
@@ -3,10 +3,10 @@
 
 #include <qp.h>
 #include "qp_lvgl.h"
-#include "qp_st7789.h"
-#include "globals.h"
-#include "native/native.h"
-#include "features/features.h"
+#include <qp_st7789.h>
+#include "ui_state.h"
+#include "event_router.h"
+#include "theme.h"
 
 static const uint16_t D_PHYSICAL_X = 172;
 static const uint16_t D_PHYSICAL_Y = 320;
@@ -15,46 +15,13 @@ static const uint16_t D_VP_OFFSET_Y = 34;
 
 static painter_device_t display;
 
-lv_color_t THEME_ACCENT;
-lv_color_t THEME_PRIMARY_BG;
-lv_color_t THEME_SECONDARY_BG;
-lv_color_t THEME_TEXT_DARK;
-lv_color_t THEME_TEXT_LIGHT;
-lv_color_t THEME_TEXT_LIGHT_GRAY;
-// UI action buttons
-lv_color_t THEME_BUTTON_1;
-lv_color_t THEME_BUTTON_2;
-lv_color_t THEME_BUTTON_3;
-lv_color_t THEME_BUTTON_4;
-
 static lv_obj_t *scr_startup;
-lv_obj_t wpm_counter_widget;
 LV_FONT_DECLARE(hooskai_font_36);
 
 
-// TODO move this to a config file optional
-void init_theme_colors(void) {
-    THEME_ACCENT = lv_color_make(59, 186, 156); // mint green
-    THEME_PRIMARY_BG = lv_color_make(46, 48, 71); // darkest blue
-    THEME_SECONDARY_BG = lv_color_make(67, 69, 92); // second darkest blue
-    THEME_TEXT_DARK = lv_color_make(112, 119, 147); // text darkest blue-gray
-    THEME_TEXT_LIGHT = lv_color_make(163, 169, 194); // text lightest blue
-    THEME_TEXT_LIGHT_GRAY = lv_color_make(180, 182, 191); // text lightest blue
-
-    // UI action buttons
-    THEME_BUTTON_1 = lv_color_make(85, 214, 79);
-    THEME_BUTTON_2 = lv_color_make(49, 72, 212);
-    THEME_BUTTON_3 = lv_color_make(242, 238, 7);
-    THEME_BUTTON_4 = lv_color_make(209, 77, 56);
-}
-
-void init_clock(lv_obj_t *screen, uint32_t time_offset) {
-    lv_obj_t *clock = lv_label_create(scr_home);
-    lv_label_set_text(clock, "12:48PM");
-    lv_obj_set_style_text_color(clock, THEME_TEXT_LIGHT, LV_PART_MAIN);
-    lv_obj_set_style_text_font(clock, &lv_font_montserrat_30, LV_PART_MAIN);
-    lv_obj_align(clock, LV_ALIGN_TOP_RIGHT, -5, 5);
-    lv_obj_fade_in(clock, 400, time_offset + 1000);
+void ui_btn_event(wui_btn_t btn) {
+    println("ui.c calling routing ------");
+    route_btn_event(btn);
 }
 
 /*
@@ -71,85 +38,7 @@ void update(void) {
 }
 */
 
-void init_ui_action_button_bar(void) {
-    lv_obj_t *action_button_bar = lv_obj_create(scr_home);
-    lv_obj_set_size(action_button_bar, 280, 30);
-    lv_obj_align(action_button_bar, LV_ALIGN_BOTTOM_MID, 0, 0);
-    lv_obj_set_scrollbar_mode(action_button_bar , LV_SCROLLBAR_MODE_OFF);
-    lv_obj_set_style_bg_color(action_button_bar, THEME_SECONDARY_BG, 0);
-    lv_obj_set_style_border_color(action_button_bar, THEME_TEXT_LIGHT, 0);
-
-    lv_obj_t *led_one = lv_obj_create(action_button_bar);
-    lv_obj_set_scrollbar_mode(led_one , LV_SCROLLBAR_MODE_OFF);
-    lv_obj_set_size(led_one , 20, 20);
-    lv_obj_set_style_bg_color(led_one , THEME_BUTTON_1, 0);
-    lv_obj_set_style_radius(led_one , LV_RADIUS_CIRCLE, 0);
-    lv_obj_t *btn_one_label = lv_label_create(action_button_bar);
-    lv_obj_set_style_text_color(btn_one_label, THEME_TEXT_LIGHT, LV_PART_MAIN);
-    lv_obj_set_style_text_font(btn_one_label, &lv_font_montserrat_14, LV_PART_MAIN);
-    lv_obj_align(led_one, LV_ALIGN_LEFT_MID, 0, 0);
-    lv_obj_align(btn_one_label, LV_ALIGN_LEFT_MID, 25, 0);
-
-    lv_obj_t *led_two = lv_obj_create(action_button_bar);
-    lv_obj_set_scrollbar_mode(led_two , LV_SCROLLBAR_MODE_OFF);
-    lv_obj_set_size(led_two , 20, 20);
-    lv_obj_set_style_bg_color(led_two , THEME_BUTTON_2, 0);
-    lv_obj_set_style_radius(led_two , LV_RADIUS_CIRCLE, 0);
-    lv_obj_t *btn_two_label = lv_label_create(action_button_bar);
-    lv_obj_set_style_text_color(btn_two_label, THEME_TEXT_LIGHT, LV_PART_MAIN);
-    lv_obj_set_style_text_font(btn_two_label, &lv_font_montserrat_14, LV_PART_MAIN);
-    lv_obj_align(led_two, LV_ALIGN_LEFT_MID, 60, 0);
-    lv_obj_align(btn_two_label, LV_ALIGN_LEFT_MID, 85, 0);
-
-    lv_obj_t *led_three = lv_obj_create(action_button_bar);
-    lv_obj_set_scrollbar_mode(led_three , LV_SCROLLBAR_MODE_OFF);
-    lv_obj_set_size(led_three, 20, 20);
-    lv_obj_set_style_bg_color(led_three, THEME_BUTTON_3, 0);
-    lv_obj_set_style_radius(led_three, LV_RADIUS_CIRCLE, 0);
-    lv_obj_t *btn_three_label = lv_label_create(action_button_bar);
-    lv_obj_set_style_text_color(btn_three_label, THEME_TEXT_LIGHT, LV_PART_MAIN);
-    lv_obj_set_style_text_font(btn_three_label, &lv_font_montserrat_14, LV_PART_MAIN);
-    lv_obj_align(led_three, LV_ALIGN_LEFT_MID, 120, 0);
-    lv_obj_align(btn_three_label, LV_ALIGN_LEFT_MID, 145, 0);
-
-    lv_obj_t *led_four = lv_obj_create(action_button_bar);
-    lv_obj_set_scrollbar_mode(led_three , LV_SCROLLBAR_MODE_OFF);
-    lv_obj_set_size(led_four, 20, 20);
-    lv_obj_set_style_bg_color(led_four, THEME_BUTTON_4, 0);
-    lv_obj_set_style_radius(led_four, LV_RADIUS_CIRCLE, 0);
-    lv_obj_t *btn_four_label = lv_label_create(action_button_bar);
-    lv_obj_set_style_text_color(btn_four_label, THEME_TEXT_LIGHT, LV_PART_MAIN);
-    lv_obj_set_style_text_font(btn_four_label, &lv_font_montserrat_14, LV_PART_MAIN);
-    lv_obj_align(led_four, LV_ALIGN_LEFT_MID, 180, 0);
-    lv_obj_align(btn_four_label, LV_ALIGN_LEFT_MID, 205, 0);
-
-}
-
-void ui_btn_event(wui_btn_t btn) {
-    println("ui.c calling routing ------");
-    btn_event_router(btn);
-}
 /*
-void ui_btn_event_one(void) {
-    init_widget_pomodoro(scr_home, THEME_TEXT_LIGHT);
-}
-
-void ui_btn_event_two(void) {
-    //pause_pomodoro();
-    init_ui_cli(scr_home);
-}
-
-void ui_btn_event_three(void) {
-    start_pomodoro();
-}
-
-void ui_btn_event_four(void) {
-    //reset_pomodoro();
-    init_widget_wpm(scr_home, THEME_TEXT_LIGHT);
-}
-void ui_encoder_switch(void) {
-    init_ui_widget_selector();
-}
 example from discord:
 typedef struct {
     painter_device_t device;
@@ -173,37 +62,6 @@ uint32_t render_gif_wpm(uint32_t trigger_time, void *cb_arg) {
     return ret ? delay_ms : 0;
 }
 */
-
-void update_layer_state_text(void) {
-    if (gbl_ui_state.active_layer_obj != NULL) {
-        lv_label_set_text_fmt(gbl_ui_state.active_layer_obj, "Layer: %s", gbl_ui_state.active_layer_text);
-    }
-}
-void update_ui_layer_state(char layer_name_text[]) {
-    strcpy(gbl_ui_state.active_layer_text, layer_name_text);
-    update_layer_state_text();
-}
-void init_ui_layer_state(void) {
-    gbl_ui_state.active_layer_obj = lv_label_create(scr_home);
-    lv_obj_set_style_text_color(gbl_ui_state.active_layer_obj, THEME_TEXT_LIGHT, LV_PART_MAIN);
-    lv_obj_set_style_text_font(gbl_ui_state.active_layer_obj, &lv_font_montserrat_14, LV_PART_MAIN);
-    lv_obj_align(gbl_ui_state.active_layer_obj, LV_ALIGN_TOP_LEFT, 10, 5);
-    update_layer_state_text();
-}
-
-
-void init_screen_home(void) {
-    scr_home = lv_obj_create(NULL);
-    lv_obj_set_style_bg_color(scr_home, THEME_PRIMARY_BG, LV_PART_MAIN);
-
-    init_clock(scr_home, 3500);
-    init_ui_action_button_bar();
-    init_ui_layer_state();
-    init_native_ui_elements();
-    init_all_wuis();
-
-    lv_scr_load_anim(scr_home, LV_SCR_LOAD_ANIM_FADE_ON, 500, 3000, false);
-}
 
 void init_screen_startup(void) {
     lv_disp_set_bg_color(lv_disp_get_default(), lv_color_black());
@@ -254,8 +112,8 @@ void init_screen_startup(void) {
 
 /* initialize the ui variables and start the UI. Runs once on startup */
 void init_ui(void) {
-    init_global_ui_state();
     init_theme_colors();
+    init_global_ui_state();
     init_screen_startup();
     init_screen_home();
 }
